Fix game_display::fake_unit assignment.
One version left the fake unit in a bad state (not on a display, but recorded
as being on one), and the other failed to keep the fake unit on a display,
even though the comments indicated it did.

With this change, the display is preserved over assignment, which is consistent
with the copy constructor and the documentation for assigning a unit to a fake.
It is not easy to preserve the fake_unit's position in the deque, and currently
no benefit in trying to, so that part of the documentation has been changed.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54978 75b3775d-3801-0410-9e12-e2e43121521d