Refactor weapon special contexts.
This allows some simplification both when the context is set and (in one
spot) when the context is used.
The semantics for the location parameters to set_specials_context() have
changed; attempts to use the old version will give a compiler error as the
function signatures have also changed (all current calls have been updated).


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55413 75b3775d-3801-0410-9e12-e2e43121521d