Correct calculation of swarm attacks when predicting a series of attacks.
(The conversion to double led to bad rounding, and the calculation should be
based on expected hit points, not original hit points.)

This is the last of the simple-to-fix bugs I detected with the stand-alone
attack_prediction / wesnoth-attack-sim combo. There is one remaining bug
(also related to swarm and a series of attacks) that might be a bear to rectify.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55440 75b3775d-3801-0410-9e12-e2e43121521d