Use size_t rather than unsigned long for the 'stored inside a pointer' type.
unsigned long is equal to void* on ILP32 (win32/linux32), LP32 (win16) and LP64 (linux64) systems, but not on LLP64 systems, which win64 is an example of.
http://www.unix.org/version2/whatsnew/lp64_wp.html
size_t is not guaranteed to be equal to void* either (see the 8086), but uintptr_t is only provided by C99 and C++11, and even then is optional.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55686 75b3775d-3801-0410-9e12-e2e43121521d