When creating [replay_start], have the snapshot overwrite the scenario config, rather than the other way around.
This is important for the RNG as the snapshot includes how often the RNG was
invoked in prestart and start evnets, while the scenario's config has the RNG
recorded as having never been invoked (naturally enough).
It might solve some other replay issues as well. (Not all replay issues are
well-documented.) 


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55801 75b3775d-3801-0410-9e12-e2e43121521d