Fix the presesent in-game command line autocompletion feature so it actually works at all
Because of the questionable console commands handler design, the command
list isn't populated until the handler's dispatch() method is called for
the first time. We can call it with an empty command string to work
around this; the empty command is ignored and we get the command list
populated so we can actually build an autocompletion dictionary with it.
Previously, the autocompletion dictionary would be always empty.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55923 75b3775d-3801-0410-9e12-e2e43121521d