gui1: Restore focus to the original textbox after resending relevant events to its input target
This effectively fixes bug #5877, *again*. r55925 introduced quite a
regression in the form of switching the focus permanently as soon as the
user presses the modifier key in Shift+<arrow> and Shift+C, requiring
improbable typing skills to get e.g. a Shift+C event _before_ a solitary
left or right Shift event.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55930 75b3775d-3801-0410-9e12-e2e43121521d