gui1: Rework textbox event forwarding from r55925 to eliminate duplicate (actually quadruplicate) events
Due to the way focusing works, the focus switch in the middle of the
forwarding action would create a duplicate event for the origin textbox,
resulting in a duplicate event for the target (a total of four events to
process).

Instead, make the implementation of gui::textbox::handle_event()
forwarding-aware, and provide a wrapper for the events engine
implementation to use by default.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55946 75b3775d-3801-0410-9e12-e2e43121521d