Rename namespace action back to "actions".
This is to avoid potential name clashes with any classses named "action", like what shadowm
found in the test suite. (The quick fix of r55968 is now reverted in favor of this fix.)
One complication is that there is already an "ai:actions" namespace, so within the "ai"
namespace, the renamed namespace is qualified as "::actions".


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@55970 75b3775d-3801-0410-9e12-e2e43121521d