Move processing of child configs from unit_type to unit_type_data::set_config().
It is a bit hacky to have unit_type modify the config passed as a parameter to its constructor,
especially when the modifications are not done by the constructor.
As for performance, this change has little impact on the net time taken by set_config(). While
it does spend more time processing configs, it takes less time building unit_types. The net
result might be a small increase in time, but any increase is comparable to the variations
in time seen amongst multiple test runs.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@56095 75b3775d-3801-0410-9e12-e2e43121521d