Avoid click dismissing moving a unit.
When click dismissing a dialogue in the past the DOWN event was used.
This lead to a bug. The obvious change was to switch to the UP event,
this lead to another bug; the dialogue was directly dismissed. Since
the game map code uses the UP and DOWN event to select a unit there is
no simple solution.

Upon entry this value stores the mouse button state at entry.  When a
button is DOWN and goes UP that button does _not_ trigger a dismissal
of the dialogue, instead that button's down state is removed from this
variable. Therefore the next UP event does dismiss the dialogue.

(Fixes bug #18970)

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@56328 75b3775d-3801-0410-9e12-e2e43121521d