display: Fix broken conditionals in display::scroll_to_tyles()
This purportedly fixes bug #18793.
(Candidate for 1.10, PLEASE TEST!)

The conditionals in question:

> if (scroll_type == ONSCREEN || ONSCREEN_WARP) {

This *always* evaluates to true when scroll_type != ONSCREEN, because
ONSCREEN_WARP is an enum member with a non-zero value. The bogus code is
effectively the same as:

> if ((scroll_type == ONSCREEN) || true) {

Which was most likely not the author's intention (fendrin in r49988).
Therefore, I've replaced the conditionals with what should be the
correct way to do what I suspect was meant above:

> if (scroll_type == ONSCREEN || scroll_type == ONSCREEN_WARP) {

This seems to fix the viewport centering issues for me, but it would be
nice if people reported their own experience with the fix before
backporting it to 1.10.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@56493 75b3775d-3801-0410-9e12-e2e43121521d